Refactor: NFT 민팅 기능 리팩토링 (#6)
* ♻️refactor[#4]: NFT 민팅 기능 리팩토링

1. 프로젝트 구조 리팩토링
2. 데이터베이스에 저장하는 컬럼 변경 (private key는 저장하지 않도록!!)
3. 멤버쉽 등록 시 트랜잭션 적용 (민팅 도중 실패하면 데이터베이스도 롤백)

* ♻️refactor[#4]: 누락된 수정사항 커밋

+++ server/Blooming/src/main/java/com/fivengers/blooming/membership/adapter/in/web/dto/NftSaleResponse.java
@@ -4,8 +4,8 @@ import com.fivengers.blooming.membership.domain.NftSale;
 import lombok.Builder;
 
 @Builder
-public record NftSaleResponse(Integer totalNftCount,
-                              Integer soldNftCount,
+public record NftSaleResponse(Long totalNftCount,
+                              Long soldNftCount,
                               Long totalNftAmount,
                               Long soldNftAmount) {
